Move classes into packages
