+++ app/src/main/java/com/enderstudy/shoppinglist/activities/PrivacyActivity.java
