+++ app/src/main/java/com/enderstudy/shoppinglist/entities/ListItem.java
